% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrapi.R
\name{aggrapi}
\alias{aggrapi}
\title{aggrapi}
\usage{
aggrapi(y)
}
\arguments{
\item{y}{\code{data.frame}}
}
\value{
\code{data.frame}
}
\description{
Function to aggregate, support function used in \code{practice_wise()} function
}
