% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_addr.R
\name{combine_addr}
\alias{combine_addr}
\title{combine_addr}
\usage{
combine_addr(path)
}
\arguments{
\item{path}{location for the \code{ADDR} files}
}
\value{
\code{dataframe}
}
\description{
Combine and convert \code{ADDR} files in the defined path to a \code{dataframe}
}
