% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COPDGene.R
\docType{data}
\name{combinedCOPDdat_RGO}
\alias{combinedCOPDdat_RGO}
\title{Chronic Obstructive Pulmonary Disease (COPD) data}
\format{
`combinedCOPDdat_RGO`
}
\usage{
data(combinedCOPDdat_RGO)
}
\description{
A subset of the COPDGene study data, obtained from the Gene Expression Omnibus database. This contains expression and clinical covariates data.
 All of the 28 genes contained in this data are identified as COPD-related by Sakornsakolpat et al. Nat Genet, 2019.
}
\examples{
data(combinedCOPDdat_RGO)
combinedCOPDdat_RGO
}
\references{
Sakornsakolpat P, Prokopenko D, Lamontagne M, et al. Genetic landscape of chronic obstructive pulmonary disease identifies heterogeneous cell-type and phenotype associations. Nat Genet. 2019;51(3):494–505.
(\href{https://pubmed.ncbi.nlm.nih.gov/30804561/}{PubMed})
Ahn S, Grimes T, Datta S. A pseudo-value regression approach for differential network analysis of co-expression data. BMC Bioinformatics, 2023;24(1):8.
(\href{https://pubmed.ncbi.nlm.nih.gov/36624383/}{PubMed})
}
\keyword{2023.}
\keyword{Ahn}
\keyword{Application}
\keyword{BMC}
\keyword{Bioinformatics.}
\keyword{al.}
\keyword{data}
\keyword{et}
\keyword{in}
\keyword{used}
