% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2dat.R
\name{csv2dat}
\alias{csv2dat}
\title{csv2dat}
\usage{
csv2dat(path)
}
\arguments{
\item{path}{location for the target .CSV file(s)}
}
\value{
\code{dataframe}
}
\description{
Combine and convert .CSV file(s) in the defined path to a \code{dataframe}
}
