% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2rds.R
\name{csv2rds}
\alias{csv2rds}
\title{csv2rds}
\usage{
csv2rds(path, output)
}
\arguments{
\item{path}{location for the target .CSV file(s)}

\item{output}{location for the output .rds file}
}
\value{
.rds file in the output location
}
\description{
Convert .CSV file(s) in the target path to a .rds file in the defined folder
}
