% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4map.R
\name{data4map}
\alias{data4map}
\title{data4map
Add the geo location IDs to the GP address files}
\usage{
data4map(x, y)
}
\arguments{
\item{x}{GP ADDR files}

\item{y}{postcode dataset}
}
\value{
a dataframe
}
\description{
data4map
Add the geo location IDs to the GP address files
}
\examples{
\dontrun{
data4map(gp_2017,full_postcode)
}
}
