% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.R
\name{EBS}
\alias{EBS}
\title{EBS}
\usage{
EBS(pvo, alpha = 0.05, B = 500, h = 1)
}
\arguments{
\item{pvo}{An object with p-values estimated from the user-provided expression data}

\item{alpha}{a level of significance (default is 0.05)}

\item{B}{size of bootstrapping (default is 500)}

\item{h}{a bandwidth (default is 1)}
}
\value{
A vector of adjusted p-values for each gene.
}
\description{
A function to calculate the adjusted p-value for each gene (Datta and Datta, 2005)
}
\references{
Datta, S and Datta, S (2005). Empirical Bayes screening of many p-values with applications to microarray studies. Bioinformatics, 21(9), 1987-1994
}
