% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalgplist.R
\name{finalgplist}
\alias{finalgplist}
\title{finalgplist}
\usage{
finalgplist(coordinate, x, y, z)
}
\arguments{
\item{coordinate}{coordinates generated from shape file}

\item{x}{polygon data generated from shape file}

\item{y}{map data generated from \code{data4map()} function}

\item{z}{GP extract data generated from GP data}
}
\value{
a dataframe
}
\description{
Generate GP lists based on the shape file and GP address
}
