% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gp_list_ccg.R
\name{generate_gp_list_ccg}
\alias{generate_gp_list_ccg}
\title{generate_gp_list_ccg}
\usage{
generate_gp_list_ccg(shape_file, region, full_postcode, gp_data)
}
\arguments{
\item{shape_file}{Shape file location}

\item{full_postcode}{Full postcode with geo information}

\item{gp_data}{GP extracted data location}
}
\value{
a character vector file with GP Practice codes
}
\description{
generate_gp_list_ccg
}
\examples{
\dontrun{
generate_gp_list_ccg(shape_file, region,full_postcode, gp_data)
}
}
