% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdmd.R
\name{importdmd}
\alias{importdmd}
\title{importdmd}
\usage{
importdmd(path)
}
\arguments{
\item{path}{dm+d excel file(s) location.}
}
\value{
A list with the following elements:
\itemize{
  \item api_map - a \code{data.frame} containing each BNF code mapped to individual API, its medicinal strength and medicinal form
  \item ing - a \code{tibble} containing all the API and its code
  \item uom - a \code{tibble} containing different unit of measurement and its code
  \item uomwdesc - a \code{data.frame} containing different unit of measurement and its code with description
  \item dform - a \code{tibble} containing different medicinal form and its code
}
}
\description{
Import dm+d files and link it to the BNF code using SNOMED mapping file
}
