% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_practice_wise.R
\name{practice_wise}
\alias{practice_wise}
\title{practice_wise}
\usage{
practice_wise(x, y, z)
}
\arguments{
\item{x}{processed prescription dataset}

\item{y}{a character vector consisting of GP practices}

\item{z}{processed BNF  mapped file}
}
\value{
multiple individual GP practice wise processed prescription dataset
}
\description{
Import NHS dataset files and generate processed individual GP practice prescription dataset
}
\examples{
\dontrun{
practice_wise(data201712_03,UK_GP2017)
}

}
