% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aracne.R
\name{run_aracne}
\alias{run_aracne}
\title{run_aracne}
\usage{
run_aracne(
  x,
  estimator = "spearman",
  disc = "none",
  nbins = NULL,
  eps = 0,
  ...
)
}
\arguments{
\item{x}{A n by p matrix of gene expression data (n samples and p genes)}

\item{estimator}{Argument is passed into minet::build.mim.}

\item{disc}{Argument is passed into minet::build.mim.}

\item{nbins}{Argument is passed into minet::build.mim.}

\item{eps}{Argument is passed into minet::aracne.}

\item{...}{Additional arguments are ignored.}
}
\value{
A p by p matrix of association scores.
}
\description{
A function to conduct co-expression analysis using ARACNE (Margolin et al. 2006). Uses the implementation from the minet package (Meyer et al. 2008).
    This function is from dnapath R package, which is archived in May 2024.
}
\references{
Margolin AA, Nemenman I, Basso K, Wiggins C, Stolovitzky G, Dalla Favera R, Califano A (2006). “ARACNE: An Algorithm for the Reconstruction of Gene Regulatory Networks in a Mammalian Cellular Context.” In BMC Bioinformatics, volume 7(1), S7. BioMed Central.

Meyer PE, Lafitte F, Bontempi G (2008). “minet: A R/Bioconductor Package for Inferring Large Transcriptional Networks using Mutual Information.” BMC Bioinformatics, 9(1), 461.
}
