% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Run in-built Shiny Apps in the package}
\usage{
runShiny(appName)
}
\arguments{
\item{appName}{In-built Shiny App name. Example: plotFlow}
}
\value{
Opens up the in-built Shiny App.
}
\description{
Run in-built Shiny Apps in the package
}
\examples{
runShiny('pda')

stpdata::runShiny('pda')

}
