% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(data, tab_name, selectdb, usernm, pwd)
}
\arguments{
\item{data}{csv data}

\item{tab_name}{Table name in the tbe DB}

\item{selectdb}{enter the interested db}

\item{usernm}{input your MySQL username}

\item{pwd}{input your MySQL password}
}
\value{
a table in MySQL
}
\description{
Upload csv to MySQL DB
}
