% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shpsubset2polygon.R
\name{shpsubset2polygon}
\alias{shpsubset2polygon}
\title{shpsubset2polygon}
\usage{
shpsubset2polygon(shapefile, region)
}
\arguments{
\item{shapefile}{shape file location}

\item{region}{region of interest}
}
\value{
a dataframe, polygon file
}
\description{
Convert shpae files to the polygons
}
\examples{
\dontrun{
shpsubset2polygon("C:/dataset/shape_files","NHS Bath and North East Somerset CCG")
}

}
