% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetahats.R
\name{thetahats}
\alias{thetahats}
\title{thetahats}
\usage{
thetahats(asso.matinput)
}
\arguments{
\item{asso.matinput}{An association matrix that is estimated from the user-provided expression data is used
as an input to compute the total connectivity of each gene.}
}
\value{
A vector containing total connectivity of each gene (i.e. continuous version of centrality measure of a network)
}
\description{
A function to compute the total connectivity of each gene from the association matrix.
}
