% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_practice_wise.R
\name{uk_practice_wise}
\alias{uk_practice_wise}
\title{uk_practice_wise}
\usage{
uk_practice_wise(x, y)
}
\arguments{
\item{x}{processed prescription dataset}

\item{y}{subset of GP practice}
}
\value{
multiple individual GP practice wise processed prescription dataset
}
\description{
uk_practice_wise
}
\examples{
\dontrun{
uk_practice_wise(data201712_03,UK_GP2017)
}

}
