% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2rds.R
\name{csv2rds}
\alias{csv2rds}
\title{csv2rds - Convert .CSV file(s) in the path defined to .rds file in a desired location}
\usage{
csv2rds(path, output)
}
\arguments{
\item{path}{.CSV file(s) location}

\item{output}{desired output .rds file with location}
}
\value{
.rds file in a desired location
}
\description{
csv2rds - Convert .CSV file(s) in the path defined to .rds file in a desired location
}
