% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdmd.R
\name{importdmd}
\alias{importdmd}
\title{importdmd}
\usage{
importdmd(path)
}
\arguments{
\item{path}{dm+d excel file(s) location. Example: importdmd ("C:/dmdDataLoader/excel/")}
}
\value{
a datatable
}
\description{
Import dm+d files and merge with June 2018 snomed mapping file.
}
