% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdmd.R
\name{importdmd}
\alias{importdmd}
\title{importdmd - import dm+d files, merge with June 2018 snomed mapping file and convert to dataframe}
\usage{
importdmd(path)
}
\arguments{
\item{path}{dm+d excel file(s) location. Example: importdmd("C:/dmdDataLoader/excel/")}
}
\value{
dataframe
}
\description{
importdmd - import dm+d files, merge with June 2018 snomed mapping file and convert to dataframe
}
