% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdmd.R
\name{importdmd}
\alias{importdmd}
\title{importdmd}
\usage{
importdmd(path)
}
\arguments{
\item{path}{dm+d excel file(s) location.}
}
\value{
a datatable
}
\description{
Import dm+d files and link it to the BNF code using June 2018 snomed mapping file
}
\examples{
\dontrun{
importdmd("C:/dmdDataLoader/excel/")
}

}
