% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdmd.R
\name{importdmd}
\alias{importdmd}
\title{importdmd}
\usage{
importdmd(path)
}
\arguments{
\item{path}{dm+d excel file(s) location.}
}
\value{
A list with the following elements:
\itemize{
  \item api_map - \code{dataframe} containing each BNF code mapped to individual API, its medicinal strength and medicinal form
  \item ing - \code{dataframe} containing all the API and its code
  \item uom - \code{dataframe} containing different unit of measurement and its code
  \item dform - \code{dataframe} containing different medicinal form and its code
}
}
\description{
Import dm+d files and link it to the BNF code using SNOMED mapping file
}
\examples{
\dontrun{
importdmd("C:/dmdDataLoader/excel/")
}

}
