% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_wise.R
\name{practice_wise}
\alias{practice_wise}
\title{practice_wise}
\usage{
practice_wise(presdata, gpvector, apimap, uom, dform, ing)
}
\arguments{
\item{presdata}{combined prescription data \code{data.frame} generated using \code{csv2dat()} function}

\item{gpvector}{\code{character} vector consisting GP practices}

\item{apimap}{BNF code mapped \code{data.frame} generated using \code{importdmd()} function}

\item{uom}{a \code{data.frame} generated using \code{importdmd()} function containing different unit of measurements and its corresponding codes}

\item{dform}{a \code{tibble} generated using \code{importdmd()} function containing different medicinal form and its corresponding codes}

\item{ing}{a \code{tibble} generated using \code{importdmd()} function containing different APIs and its corresponding codes}
}
\value{
multiple individual GP practice wise processed prescription dataset
}
\description{
Import NHS dataset files and generate processed individual GP practice prescription dataset
}
\examples{
\dontrun{
 
## an example character vector containing different GP Practices
gp_practice <- c("L81069","L81070","L81071","L81072")

## Define the files output location
setwd("C:/Datasets/practice_wise/2018")

## data201812 - generated using the csv2dat() function

## modify objects generated using the impordmd() function 

## api mapping 
api_map <- dmdfile$api_map

## unit of measurement  
uom <- dmdfile$uomwdesc \%>\%
        dplyr::rename(UOM = CD)

## medicinal form 
dform <- dmdfile$dform

## ingredients
ing <- dmdfile$ing \%>\%
        dplyr::rename(API_CODE = 1)

## execute practice_wise function to generate multiple files
practice_wise(data201812, gp_practice, api_map, uom, dform, ing)

}

}
