% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(data, tab_name, selectdb)
}
\arguments{
\item{data}{csv data}

\item{tab_name}{Table name in the tbe DB}
}
\value{

}
\description{
Upload csv to MySQL DB
}
